weather message request weather images anecdote encoding fix some modifications

+++ src/com/frontier/botChat/utils/Parse/WeatherTemp.java
@@ -3,13 +3,13 @@ package com.frontier.botChat.utils.Parse;
 public class WeatherTemp {
 
     private double temp;
-    private int humidity;
+    private double humidity;
 
     public double getTemp() {
         return temp;
     }
 
-    public int getHumidity() {
+    public double getHumidity() {
         return humidity;
     }
 }
